+++ src/llama-context.h
@@ -219,3 +219,7 @@ static void llama_output_reorder(struct llama_context * ctx) {
         out_ids.clear();
     }
 }
+
+// For internal test use
+// TODO: remove
+const std::vector<std::pair<std::string, struct ggml_tensor *>> & llama_internal_get_tensor_map(struct llama_context * ctx);
